kjhkjhhj/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==1.0.0
gunicorn==20.1.0
Jinja2==3.1.2
bcrypt==4.0.1
PyJWT==2.6.0
python-decouple==3.8.0
email-validator
